requirements.txt
azure.ai.contentsafety
azure-cognitiveservices-vision-customvision
python-multipart
pydantic
fastapi